requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.7
openai>=1.0.0
pdfplumber>=0.10.0
chromadb>=0.3.29
tqdm>=4.66.0
python-dotenv>=1.0.0
